app/requirements.txt
RPi.GPIO
picamera2
numpy
pytesseract
